Added main in root
